add DynamicStructOfStaticStruct
